api/requirements.txt
fastapi
uvicorn
motor
pydantic
pyspark
pytesseract
pillow
numpy
dotenv
passlib
PyJWT
python-multipart
bcrypt
requests
pymupdf
openai